web_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
SQLAlchemy
passlib
python-multipart
jinja2
